Update README.md
There are so many books one can find on infosec, from Hacking Exposed to the Web Application Handbook. However, the books that I suggest in this pull request are of a different kind. They portray real life hacking scenarios where the reader shadows a hacker trying to break into a company. They get to experience the frustraiton, joy and excitment of a real hacking engagement.
Furthermore, the techniques and tips focus on common systems and network configurations: Windows, Mainframes, Active Directory, Linux, etc.